Features, improvements, and bug fixes - v1.1.1(290)

+++ Packages/RuuviService/Sources/RuuviServiceAuth/RuuviServiceAuthImpl.swift
@@ -35,6 +35,10 @@ public final class RuuviServiceAuthImpl: RuuviServiceAuth {
         storage.readAll()
             .on(success: { [weak self] localSensors in
                 guard let sSelf = self else { return }
+                guard localSensors.count != 0 else {
+                    promise.succeed(value: true)
+                    return
+                }
                 localSensors.filter({ $0.isClaimed || $0.isCloud }).forEach { sensor in
                     let deleteSensorOperation = sSelf.pool.delete(sensor)
                     let deleteRecordsOperation = sSelf.pool.deleteAllRecords(sensor.id)
